+++ src/main/java/net/fullstack7/edusecond/edusecond/domain/member/MemberVO.java
@@ -1,31 +1,25 @@
-package net.fullstack7.edusecond.edusecond.dto;
+package net.fullstack7.edusecond.edusecond.domain.member;
 
 import java.time.LocalDate;
-import java.util.List;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 
 @Getter
-@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MemberDTO {
+public class MemberVO {
     private String userId;
+    private String userPw;
     private String userName;
     private String userEmail;
     private String userPhone;
     private String userAddress;
     private String userPostcode;
     private LocalDate userBirth;
-    
-    // 추가 필드
-    private int orderCount;
-    private int reviewCount;
-    private double avgRating;
-    private List<ProductDTO> wishlist;
+    private LocalDateTime regDate;
+    private LocalDateTime modifyDate;
 }
