+++ src/main/java/net/fullstack7/edusecond/edusecond/dto/member/MemberDTO.java
@@ -1,25 +1,32 @@
-package net.fullstack7.edusecond.edusecond.domain;
+package net.fullstack7.edusecond.edusecond.dto.member;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
+import net.fullstack7.edusecond.edusecond.dto.product.ProductDTO;
+
 
 @Getter
+@Setter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MemberVO {
+public class MemberDTO {
     private String userId;
-    private String userPw;
     private String userName;
     private String userEmail;
     private String userPhone;
     private String userAddress;
     private String userPostcode;
     private LocalDate userBirth;
-    private LocalDateTime regDate;
-    private LocalDateTime modifyDate;
+    
+    // 추가 필드
+    private int orderCount;
+    private int reviewCount;
+    private double avgRating;
+    private List<ProductDTO> wishlist;
 }
