+++ src/main/java/net/fullstack7/edusecond/edusecond/mapper/MemberMapper.java
@@ -1,6 +1,6 @@
 package net.fullstack7.edusecond.edusecond.mapper;
 
-import net.fullstack7.edusecond.edusecond.domain.MemberVO;
+import net.fullstack7.edusecond.edusecond.domain.member.MemberVO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
