+++ src/test/java/net/fullstack7/edusecond/edusecond/ProductServiceTests.java
@@ -1,19 +1,13 @@
 package net.fullstack7.edusecond.edusecond;
 
 import lombok.extern.log4j.Log4j2;
-import net.fullstack7.edusecond.edusecond.domain.MemberVO;
-import net.fullstack7.edusecond.edusecond.domain.ProductVO;
-import net.fullstack7.edusecond.edusecond.dto.ProductDTO;
-import net.fullstack7.edusecond.edusecond.service.ProductServiceIf;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
+import net.fullstack7.edusecond.edusecond.dto.product.ProductDTO;
+import net.fullstack7.edusecond.edusecond.service.product.ProductServiceIf;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
